add some code comments
